Introduce basic block level parsing for math equations
This change is necessary to ensure that when parsing markdown containing
mathematical equations. These are preserved and rendered correctly.

This change addresses the need by bringing in a regular expression to
the parser that can match math blocks surrounded by $$ but not when
those blocks are inside a fenced code block.

Additionally, in order to ensure that the Latex markup is maintained, it
escapes all backslashes in the text.

As a result of this change, in local testing the following side effects
were observed:

Whilst I've not noticed any detrimental side-effects, adding block math
code to the markdown does require that it is preceded by 2 newlines.

This would be normal flow as it's text EOL + 1 blank line.